# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AnalysisInputFamily = typeof AnalysisInputFamily[keyof typeof AnalysisInputFamily];


export const AnalysisInputFamily = {
  matches: 'matches',
  'even-odd': 'even-odd',
  'over-under': 'over-under',
  'rise-fall': 'rise-fall',
  all: 'all',
} as const;

export interface AnalysisInput {
  /** @maxLength 64 */
  symbol?: string;
  /**
     * @minItems 5
     * @maxItems 2000
     * @items.exclusiveMinimum 0
     */
  ticks: number[];
  family: AnalysisInputFamily;
  /**
     * Optional legacy override. The dashboard leaves this unset so the analysis selects the barrier.
     * @minimum 1
     * @maximum 8
     */
  barrier?: number;
}

export interface CandleBar {
  /** @minimum 0 */
  epoch: number;
  /** @exclusiveMinimum 0 */
  open: number;
  /** @exclusiveMinimum 0 */
  high: number;
  /** @exclusiveMinimum 0 */
  low: number;
  /** @exclusiveMinimum 0 */
  close: number;
}

export interface CandleOpenAnalysisInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  symbol: string;
  intervalSeconds: 60;
  /** @minimum 0 */
  candleEpoch: number;
  /** @exclusiveMinimum 0 */
  openPrice: number;
  /**
     * @minItems 30
     * @maxItems 500
     */
  completedCandles: CandleBar[];
}

export interface CandleOpenFeatures {
  emaSpreadAtr: number;
  rsi: number;
  adx: number;
  plusDi: number;
  minusDi: number;
  atr: number;
  bollingerPosition: number;
  returnAutocorrelation: number;
  volatilityClustering: number;
  efficiencyRatio: number;
  openingGapAtr: number;
  upCandleRate: number;
  trendTStatistic: number;
}

export type CandleOpenForecastResponseLabel = typeof CandleOpenForecastResponseLabel[keyof typeof CandleOpenForecastResponseLabel];


export const CandleOpenForecastResponseLabel = {
  RISE: 'RISE',
  FALL: 'FALL',
} as const;

export interface AnalysisEngineSummary {
  name: string;
  description: string;
  score: number;
  signal: string;
  detail: string;
}

export interface CandleOpenForecastResponse {
  symbol: string;
  intervalSeconds: number;
  candleEpoch: number;
  openPrice: number;
  label: CandleOpenForecastResponseLabel;
  /**
     * @minimum 0.5
     * @maximum 0.99
     */
  probability: number;
  /**
     * @minimum 50
     * @maximum 94
     */
  confidence: number;
  sampleSize: number;
  effectiveSampleSize: number;
  detail: string;
  /** @minItems 1 */
  engines: AnalysisEngineSummary[];
  features: CandleOpenFeatures;
  methodNote: string;
}

export interface AnalysisFactor {
  name: string;
  value: number;
  note: string;
  engine: string;
}

export type AnalysisPredictionTone = typeof AnalysisPredictionTone[keyof typeof AnalysisPredictionTone];


export const AnalysisPredictionTone = {
  mint: 'mint',
  amber: 'amber',
  blue: 'blue',
} as const;

export interface AnalysisPrediction {
  family: string;
  label: string;
  probability: number;
  confidence: number;
  direction: string;
  detail: string;
  tone: AnalysisPredictionTone;
}

export type AnalysisSignalTone = typeof AnalysisSignalTone[keyof typeof AnalysisSignalTone];


export const AnalysisSignalTone = {
  mint: 'mint',
  amber: 'amber',
  blue: 'blue',
} as const;

export interface AnalysisSignal {
  label: string;
  direction: string;
  probability: number;
  confidence: number;
  detail: string;
  tone: AnalysisSignalTone;
}

export type AnalysisMetricsVolatilityRegime = typeof AnalysisMetricsVolatilityRegime[keyof typeof AnalysisMetricsVolatilityRegime];


export const AnalysisMetricsVolatilityRegime = {
  compressed: 'compressed',
  balanced: 'balanced',
  expanded: 'expanded',
} as const;

export interface AnalysisMetrics {
  mean: number;
  median: number;
  stdDev: number;
  coefficientVariation: number;
  min: number;
  max: number;
  range: number;
  skewness: number;
  kurtosis: number;
  zScore: number;
  slope: number;
  returnMean: number;
  returnStdDev: number;
  emaFast: number;
  emaSlow: number;
  rsi: number;
  atrProxy: number;
  bollingerPosition: number;
  lag1Autocorrelation: number;
  shannonEntropy: number;
  normalizedEntropy: number;
  chiSquare: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  chiSquarePValue: number;
  runCount: number;
  longestRun: number;
  positiveRate: number;
  /**
     * @minItems 10
     * @maxItems 10
     */
  lastDigitDistribution: number[];
  /**
     * @minItems 10
     * @maxItems 10
     * @items.minItems 10
     * @items.maxItems 10
     */
  transitionMatrix: number[][];
  /**
     * @minItems 10
     * @maxItems 10
     */
  transitionFromLast: number[];
  /**
     * @minItems 10
     * @maxItems 10
     */
  nextDigitProbabilities: number[];
  volatilityRegime: AnalysisMetricsVolatilityRegime;
  /**
     * @minItems 10
     * @maxItems 10
     * @items.minimum 0
     */
  digitCounts: number[];
  /**
     * @minItems 10
     * @maxItems 10
     * @items.minItems 10
     * @items.maxItems 10
     * @items.items.minimum 0
     */
  transitionCounts: number[][];
  /**
     * @minItems 5
     * @maxItems 5
     * @items.minimum -1
     * @items.maximum 1
     */
  autocorrelationByLag: number[];
  /** @minimum 0 */
  ljungBoxQ: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  ljungBoxPValue: number;
  /** @minimum 0 */
  jarqueBera: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  normalityPValue: number;
  runsZScore: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  runsPValue: number;
  trendTStatistic: number;
  /**
     * @minimum -1
     * @maximum 1
     */
  volatilityAutocorrelation: number;
  structuralShiftZScore: number;
  valueAtRisk95: number;
  expectedShortfall95: number;
  /** @minimum 0 */
  maxDrawdown: number;
  /** @minimum 0 */
  effectiveSampleSize: number;
}

export interface AnalysisResponse {
  version: string;
  symbol: string;
  sampleSize: number;
  engines: AnalysisEngineSummary[];
  signal: AnalysisSignal;
  /** @minItems 4 */
  predictions: AnalysisPrediction[];
  /** @minItems 4 */
  factors: AnalysisFactor[];
  metrics: AnalysisMetrics;
  methodNote: string;
}