# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Tick analysis and prediction engines
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analysis:
    post:
      operationId: analyzeTicks
      tags: [analysis]
      summary: Analyze a tick window
      description: Runs statistical, transition, digit, momentum, and volatility engines over a tick window.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"
        "400":
          description: Invalid tick window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/candle-open:
    post:
      operationId: analyzeCandleOpen
      tags: [analysis]
      summary: Predict direction at the open of a one-minute candle
      description: Uses completed one-minute candles and the current candle open to estimate whether the current candle will close higher or lower. The current candle's later prices are never included.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandleOpenAnalysisInput"
      responses:
        "200":
          description: Candle-open forecast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleOpenForecastResponse"
        "400":
          description: Invalid candle history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AnalysisInput:
      type: object
      required: [ticks, family]
      properties:
        symbol:
          type: string
          maxLength: 64
        ticks:
          type: array
          minItems: 5
          maxItems: 2000
          items:
            type: number
            exclusiveMinimum: 0
        family:
          type: string
          enum: [matches, even-odd, over-under, rise-fall, all]
        barrier:
          type: integer
          minimum: 1
          maximum: 8
          description: Optional legacy override. The dashboard leaves this unset so the analysis selects the barrier.
    CandleBar:
      type: object
      required: [epoch, open, high, low, close]
      properties:
        epoch: { type: integer, minimum: 0 }
        open: { type: number, exclusiveMinimum: 0 }
        high: { type: number, exclusiveMinimum: 0 }
        low: { type: number, exclusiveMinimum: 0 }
        close: { type: number, exclusiveMinimum: 0 }
    CandleOpenAnalysisInput:
      type: object
      required: [symbol, intervalSeconds, candleEpoch, openPrice, completedCandles]
      properties:
        symbol:
          type: string
          minLength: 2
          maxLength: 30
        intervalSeconds:
          type: integer
          const: 60
        candleEpoch:
          type: integer
          minimum: 0
        openPrice:
          type: number
          exclusiveMinimum: 0
        completedCandles:
          type: array
          minItems: 30
          maxItems: 500
          items:
            $ref: "#/components/schemas/CandleBar"
    CandleOpenFeatures:
      type: object
      required: [emaSpreadAtr, rsi, adx, plusDi, minusDi, atr, bollingerPosition, returnAutocorrelation, volatilityClustering, efficiencyRatio, openingGapAtr, upCandleRate, trendTStatistic]
      properties:
        emaSpreadAtr: { type: number }
        rsi: { type: number }
        adx: { type: number }
        plusDi: { type: number }
        minusDi: { type: number }
        atr: { type: number }
        bollingerPosition: { type: number }
        returnAutocorrelation: { type: number }
        volatilityClustering: { type: number }
        efficiencyRatio: { type: number }
        openingGapAtr: { type: number }
        upCandleRate: { type: number }
        trendTStatistic: { type: number }
    CandleOpenForecastResponse:
      type: object
      required: [symbol, intervalSeconds, candleEpoch, openPrice, label, probability, confidence, sampleSize, effectiveSampleSize, detail, engines, features, methodNote]
      properties:
        symbol: { type: string }
        intervalSeconds: { type: integer }
        candleEpoch: { type: integer }
        openPrice: { type: number }
        label:
          type: string
          enum: [RISE, FALL]
        probability:
          type: number
          minimum: 0.5
          maximum: 0.99
        confidence:
          type: integer
          minimum: 50
          maximum: 94
        sampleSize: { type: integer }
        effectiveSampleSize: { type: number }
        detail: { type: string }
        engines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AnalysisEngineSummary"
        features:
          $ref: "#/components/schemas/CandleOpenFeatures"
        methodNote: { type: string }
    AnalysisEngineSummary:
      type: object
      required: [name, description, score, signal, detail]
      properties:
        name:
          type: string
        description:
          type: string
        score:
          type: number
        signal:
          type: string
        detail:
          type: string
    AnalysisFactor:
      type: object
      required: [name, value, note, engine]
      properties:
        name:
          type: string
        value:
          type: number
        note:
          type: string
        engine:
          type: string
    AnalysisPrediction:
      type: object
      required: [family, label, probability, confidence, direction, detail, tone]
      properties:
        family:
          type: string
        label:
          type: string
        probability:
          type: number
        confidence:
          type: number
        direction:
          type: string
        detail:
          type: string
        tone:
          type: string
          enum: [mint, amber, blue]
    AnalysisSignal:
      type: object
      required: [label, direction, probability, confidence, detail, tone]
      properties:
        label:
          type: string
        direction:
          type: string
        probability:
          type: number
        confidence:
          type: number
        detail:
          type: string
        tone:
          type: string
          enum: [mint, amber, blue]
    AnalysisMetrics:
      type: object
      required:
        - mean
        - median
        - stdDev
        - coefficientVariation
        - min
        - max
        - range
        - skewness
        - kurtosis
        - zScore
        - slope
        - returnMean
        - returnStdDev
        - emaFast
        - emaSlow
        - rsi
        - atrProxy
        - bollingerPosition
        - lag1Autocorrelation
        - shannonEntropy
        - normalizedEntropy
        - chiSquare
        - chiSquarePValue
        - runCount
        - longestRun
        - positiveRate
        - lastDigitDistribution
        - transitionMatrix
        - transitionFromLast
        - nextDigitProbabilities
        - volatilityRegime
        - digitCounts
        - transitionCounts
        - autocorrelationByLag
        - ljungBoxQ
        - ljungBoxPValue
        - jarqueBera
        - normalityPValue
        - runsZScore
        - runsPValue
        - trendTStatistic
        - volatilityAutocorrelation
        - structuralShiftZScore
        - valueAtRisk95
        - expectedShortfall95
        - maxDrawdown
        - effectiveSampleSize
      properties:
        mean: { type: number }
        median: { type: number }
        stdDev: { type: number }
        coefficientVariation: { type: number }
        min: { type: number }
        max: { type: number }
        range: { type: number }
        skewness: { type: number }
        kurtosis: { type: number }
        zScore: { type: number }
        slope: { type: number }
        returnMean: { type: number }
        returnStdDev: { type: number }
        emaFast: { type: number }
        emaSlow: { type: number }
        rsi: { type: number }
        atrProxy: { type: number }
        bollingerPosition: { type: number }
        lag1Autocorrelation: { type: number }
        shannonEntropy: { type: number }
        normalizedEntropy: { type: number }
        chiSquare: { type: number }
        chiSquarePValue: { type: number, minimum: 0, maximum: 1 }
        runCount: { type: integer }
        longestRun: { type: integer }
        positiveRate: { type: number }
        lastDigitDistribution:
          type: array
          minItems: 10
          maxItems: 10
          items: { type: number }
        transitionMatrix:
          type: array
          minItems: 10
          maxItems: 10
          items:
            type: array
            minItems: 10
            maxItems: 10
            items: { type: number }
        transitionFromLast:
          type: array
          minItems: 10
          maxItems: 10
          items: { type: number }
        nextDigitProbabilities:
          type: array
          minItems: 10
          maxItems: 10
          items: { type: number }
        volatilityRegime:
          type: string
          enum: [compressed, balanced, expanded]
        digitCounts:
          type: array
          minItems: 10
          maxItems: 10
          items: { type: integer, minimum: 0 }
        transitionCounts:
          type: array
          minItems: 10
          maxItems: 10
          items:
            type: array
            minItems: 10
            maxItems: 10
            items: { type: integer, minimum: 0 }
        autocorrelationByLag:
          type: array
          minItems: 5
          maxItems: 5
          items: { type: number, minimum: -1, maximum: 1 }
        ljungBoxQ: { type: number, minimum: 0 }
        ljungBoxPValue: { type: number, minimum: 0, maximum: 1 }
        jarqueBera: { type: number, minimum: 0 }
        normalityPValue: { type: number, minimum: 0, maximum: 1 }
        runsZScore: { type: number }
        runsPValue: { type: number, minimum: 0, maximum: 1 }
        trendTStatistic: { type: number }
        volatilityAutocorrelation: { type: number, minimum: -1, maximum: 1 }
        structuralShiftZScore: { type: number }
        valueAtRisk95: { type: number }
        expectedShortfall95: { type: number }
        maxDrawdown: { type: number, minimum: 0 }
        effectiveSampleSize: { type: number, minimum: 0 }
    AnalysisResponse:
      type: object
      required: [version, symbol, sampleSize, engines, signal, predictions, factors, metrics, methodNote]
      properties:
        version:
          type: string
        symbol:
          type: string
        sampleSize:
          type: integer
        engines:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisEngineSummary"
        signal:
          $ref: "#/components/schemas/AnalysisSignal"
        predictions:
          type: array
          minItems: 4
          items:
            $ref: "#/components/schemas/AnalysisPrediction"
        factors:
          type: array
          minItems: 4
          items:
            $ref: "#/components/schemas/AnalysisFactor"
        metrics:
          $ref: "#/components/schemas/AnalysisMetrics"
        methodNote:
          type: string