# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisInputFamily } from './analysisInputFamily';

export interface AnalysisInput {
  /** @maxLength 64 */
  symbol?: string;
  /**
     * @minItems 5
     * @maxItems 2000
     * @items.exclusiveMinimum 0
     */
  ticks: number[];
  family: AnalysisInputFamily;
  /**
     * Optional legacy override. The dashboard leaves this unset so the analysis selects the barrier.
     * @minimum 1
     * @maximum 8
     */
  barrier?: number;
}