# lib/api-zod/src/generated/types/analysisMetrics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisMetricsVolatilityRegime } from './analysisMetricsVolatilityRegime';

export interface AnalysisMetrics {
  mean: number;
  median: number;
  stdDev: number;
  coefficientVariation: number;
  min: number;
  max: number;
  range: number;
  skewness: number;
  kurtosis: number;
  zScore: number;
  slope: number;
  returnMean: number;
  returnStdDev: number;
  emaFast: number;
  emaSlow: number;
  rsi: number;
  atrProxy: number;
  bollingerPosition: number;
  lag1Autocorrelation: number;
  shannonEntropy: number;
  normalizedEntropy: number;
  chiSquare: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  chiSquarePValue: number;
  runCount: number;
  longestRun: number;
  positiveRate: number;
  /**
     * @minItems 10
     * @maxItems 10
     */
  lastDigitDistribution: number[];
  /**
     * @minItems 10
     * @maxItems 10
     * @items.minItems 10
     * @items.maxItems 10
     */
  transitionMatrix: number[][];
  /**
     * @minItems 10
     * @maxItems 10
     */
  transitionFromLast: number[];
  /**
     * @minItems 10
     * @maxItems 10
     */
  nextDigitProbabilities: number[];
  volatilityRegime: AnalysisMetricsVolatilityRegime;
  /**
     * @minItems 10
     * @maxItems 10
     * @items.minimum 0
     */
  digitCounts: number[];
  /**
     * @minItems 10
     * @maxItems 10
     * @items.minItems 10
     * @items.maxItems 10
     * @items.items.minimum 0
     */
  transitionCounts: number[][];
  /**
     * @minItems 5
     * @maxItems 5
     * @items.minimum -1
     * @items.maximum 1
     */
  autocorrelationByLag: number[];
  /** @minimum 0 */
  ljungBoxQ: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  ljungBoxPValue: number;
  /** @minimum 0 */
  jarqueBera: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  normalityPValue: number;
  runsZScore: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  runsPValue: number;
  trendTStatistic: number;
  /**
     * @minimum -1
     * @maximum 1
     */
  volatilityAutocorrelation: number;
  structuralShiftZScore: number;
  valueAtRisk95: number;
  expectedShortfall95: number;
  /** @minimum 0 */
  maxDrawdown: number;
  /** @minimum 0 */
  effectiveSampleSize: number;
}