# lib/api-zod/src/generated/types/candleBar.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CandleBar {
  /** @minimum 0 */
  epoch: number;
  /** @exclusiveMinimum 0 */
  open: number;
  /** @exclusiveMinimum 0 */
  high: number;
  /** @exclusiveMinimum 0 */
  low: number;
  /** @exclusiveMinimum 0 */
  close: number;
}