# lib/api-zod/src/generated/types/candleOpenAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CandleBar } from './candleBar';

export interface CandleOpenAnalysisInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  symbol: string;
  intervalSeconds: 60;
  /** @minimum 0 */
  candleEpoch: number;
  /** @exclusiveMinimum 0 */
  openPrice: number;
  /**
     * @minItems 30
     * @maxItems 500
     */
  completedCandles: CandleBar[];
}