# lib/api-zod/src/generated/types/candleOpenFeatures.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CandleOpenFeatures {
  emaSpreadAtr: number;
  rsi: number;
  adx: number;
  plusDi: number;
  minusDi: number;
  atr: number;
  bollingerPosition: number;
  returnAutocorrelation: number;
  volatilityClustering: number;
  efficiencyRatio: number;
  openingGapAtr: number;
  upCandleRate: number;
  trendTStatistic: number;
}