# lib/api-zod/src/generated/types/candleOpenForecastResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisEngineSummary } from './analysisEngineSummary';
import type { CandleOpenFeatures } from './candleOpenFeatures';
import type { CandleOpenForecastResponseLabel } from './candleOpenForecastResponseLabel';

export interface CandleOpenForecastResponse {
  symbol: string;
  intervalSeconds: number;
  candleEpoch: number;
  openPrice: number;
  label: CandleOpenForecastResponseLabel;
  /**
     * @minimum 0.5
     * @maximum 0.99
     */
  probability: number;
  /**
     * @minimum 50
     * @maximum 94
     */
  confidence: number;
  sampleSize: number;
  effectiveSampleSize: number;
  detail: string;
  /** @minItems 1 */
  engines: AnalysisEngineSummary[];
  features: CandleOpenFeatures;
  methodNote: string;
}